mountoperation: Silence gcc
Also, printf()ing a GPid should be forbidden.

+++ gtk/gtkmountoperation.c
@@ -1090,7 +1090,7 @@ add_pid_to_process_list_store (GtkMountOperation              *mount_operation,
                                     &pixbuf);
 
   if (name == NULL)
-    name = g_strdup_printf (_("Unknown Application (PID %d)"), (int) pid);
+    name = g_strdup_printf (_("Unknown Application (PID %d)"), (int) (gssize) pid);
 
   if (command_line == NULL)
     command_line = g_strdup ("");
